order and cart

+++ app/category/controller.js
@@ -1,8 +1,22 @@
 const Category = require('./model');
 
+const {policyFor} = require('../policy/index');
+
+
 async function store(req, res, next) {
     let payload = req.body;
     try {
+
+        //TODO cek policy
+        let policy = policyFor(req.user);
+
+        if(!policy.can('create', 'Category')){
+            return res.json({
+                error:1,
+                message:'Anda tidak memiliki akses untuk membuat kategori'
+            });
+        }
+
         let category = new Category(payload);
 
         await category.save();
@@ -27,6 +41,17 @@ async function update(req, res, next){
     let payload = req.body;
 
     try {
+
+        //TODO cek policy
+        let policy = policyFor(req.user);
+
+        if(!policy.can('update', 'Category')){
+            return res.json({
+                error:1,
+                message: 'Anda tidak memiliki akses untuk mengupdate kategori'
+            })
+        }
+
         let category = await Category.findOneAndUpdate({_id: req.params.id},payload,{
             new: true, runValidators: true
         })
@@ -51,6 +76,17 @@ async function update(req, res, next){
 
 async function destroy(req, res, next) {
     try {
+
+        //TODO cek policy
+        let policy = policyFor(req.user);
+
+        if(!policy.can('delete', 'Category')){
+            return res.json({
+                error:1,
+                message:'Anda tidak memiliki akses untuk menghapus kategori'
+            })
+        }
+
         const deleted = await Category.findOneAndDelete({_id: req.params.id});
 
         return res.json(deleted);
